# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/products":{"get":{"tags":["product-controller"],"summary":"Get all products","description":"Retrieve a list of all products.","operationId":"getAllProducts","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductVO"}}}}},"200":{"description":"Successfully retrieved the list of products","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductVO"}}}}}}},"post":{"tags":["product-controller"],"summary":"Create a new product","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductVO"}}},"required":true},"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"201":{"description":"Product created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductVO"}}}},"400":{"description":"Invalid input","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductVO"}}}}}}},"/products/{id}":{"get":{"tags":["product-controller"],"summary":"Retrieve details for a specific product by ID","operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Product not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Found the product","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductVO"}}}}}},"delete":{"tags":["product-controller"],"summary":"Delete a product by ID","operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Product not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"204":{"description":"Product deleted"}}},"patch":{"tags":["product-controller"],"summary":"Update an existing product","operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductVO"}}},"required":true},"responses":{"404":{"description":"Product not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"200":{"description":"Product updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductVO"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"message":{"type":"string"},"path":{"type":"string"}}},"ProductVO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"image":{"type":"string"},"category":{"type":"string"},"price":{"type":"number","format":"double"},"quantity":{"type":"integer","format":"int32"},"internalReference":{"type":"string"},"shellId":{"type":"integer","format":"int64"},"inventoryStatus":{"type":"string","enum":["INSTOCK","LOWSTOCK","OUTOFSTOCK"]},"rating":{"type":"number","format":"double"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}}